Factors out behavior from ActiveFedora::Rdf::ObjectResource to mixin ActiveFedora::Rdf::Persistence.
The change allows any ActiveTriples::Resource descendant which includes the mixin
to be used as the resource class for an ActiveFedora::RDFDatastream, rather than
requiring an ObjectResource subclass.
