+++ lib/active_fedora/rdf/rdf_datastream.rb
@@ -18,6 +18,9 @@ module ActiveFedora
 
       # Utility method which can be overridden to determine the object
       # resource that is created.
+      # 
+      # @return [Class] the object resource class
+      #   Must be a descendant of ActiveTriples::Resource and include ActiveFedora::Rdf::Persistence.
       def resource_class
         Rdf::ObjectResource
       end
